added bbox/centers and moved to support.grid
